refactor(tests): split fake repository jobs
